Narf

+++ src/lib/expression/exists_expression.cpp
@@ -51,7 +51,7 @@ bool ExistsExpression::_shallow_equals(const AbstractExpression& expression) con
 }
 
 size_t ExistsExpression::_shallow_hash() const {
-  return exists_expression_type == ExistsExpressionType::Exists;
+  return static_cast<size_t>(exists_expression_type == ExistsExpressionType::Exists);
 }
 
 bool ExistsExpression::_on_is_nullable_on_lqp(const AbstractLQPNode& /*lqp*/) const {
